etl/requirements.txt
psycopg==3.2.3
python-dotenv==1.0.1
python-dateutil==2.9.0
loguru==0.7.2
pydantic==2.9.2
backoff==2.2.1
redis==5.2.0